requirements.txt
# Core dependencies
openai==1.12.0
git+https://github.com/openai/whisper.git
gTTS==2.5.0

# Audio processing
pyaudio==0.2.14
sounddevice==0.4.6
soundfile==0.12.1
numpy==1.24.3

# Environment and utilities
python-dotenv==1.0.0
pydub==0.25.1

# Optional: for notebook support
jupyter==1.0.0
ipython==8.12.0

# Development
pytest==7.4.0
black==23.7.0
flake8==6.1.0
